+++ tests/test_contrib.py
@@ -29,6 +29,7 @@ def test_minimum_env(monkeypatch):
             require_browser_auth=False,
             disable_push=False,
             unauthenticated_push=False,
+            scip_policy="none",
         ),
     )
 
@@ -45,6 +46,7 @@ def test_complete_env(monkeypatch):
             "KLAUS_REQUIRE_BROWSER_AUTH": "1",
             "KLAUS_DISABLE_PUSH": "false",
             "KLAUS_UNAUTHENTICATED_PUSH": "0",
+            "KLAUS_SCIP_POLICY": "ALL",
         },
         ([TEST_REPO_NO_NAMESPACE], TEST_SITE_NAME),
         dict(
@@ -53,6 +55,7 @@ def test_complete_env(monkeypatch):
             require_browser_auth=True,
             disable_push=False,
             unauthenticated_push=False,
+            scip_policy="ALL",
         ),
     )